+++ eslint.config.mjs
@@ -70,6 +70,7 @@ export default [
       // Filtering tsc output files (i.e. if there a foo.ts, we ignore foo.js):
       (f, _, files) => f.endsWith('js') && files.includes(f.replace(/(\.[cm]?)js$/, '$1ts')),
     ),
+    '!test/fixtures/test-inspector-dom-storage.mjs',
     '!test/fixtures/test-runner',
     'test/fixtures/test-runner/*',
     '!test/fixtures/test-runner/output',
